Fixed image topics not having all the same stamp
As `mFrameTimestamp` is updated in a parallel thread, `mFrameTimestamp` may change before the end of publishing images loop, so some images published at the same time won't get the same stamp. This is a problem if we have a subscriber using exact time synchronization to get RGB+Depth with exactly the same stamp, sometimes they match, sometimes they not because of the threading race on `mFrameTimestamp` member.